requirements.txt
## REDDilbert Requirements
pyyaml
bs4
praw
